+++ lib/Service/GoogleDriveAPIService.php
@@ -11,12 +11,15 @@
 
 namespace OCA\Google\Service;
 
-use OCP\IL10N;
+use Datetime;
+use OCP\Files\Folder;
+use OCP\Files\InvalidPathException;
+use OCP\Files\NotPermittedException;
 use OCP\IConfig;
 use OCP\Files\IRootFolder;
 use OCP\Files\FileInfo;
-use OCP\Files\Node;
 use OCP\BackgroundJob\IJobList;
+use OCP\Lock\ILockingProvider;
 use Psr\Log\LoggerInterface;
 use OCP\Files\NotFoundException;
 use OCP\Lock\LockedException;
@@ -25,27 +28,58 @@ use OCA\Google\AppInfo\Application;
 use OCA\Google\BackgroundJob\ImportDriveJob;
 
 class GoogleDriveAPIService {
-
-	private $l10n;
+	/**
+	 * @var string
+	 */
+	private $appName;
+	/**
+	 * @var LoggerInterface
+	 */
 	private $logger;
+	/**
+	 * @var IConfig
+	 */
+	private $config;
+	/**
+	 * @var IRootFolder
+	 */
+	private $root;
+	/**
+	 * @var IJobList
+	 */
+	private $jobList;
+	/**
+	 * @var GoogleAPIService
+	 */
+	private $googleApiService;
+	/**
+	 * @var UserScopeService
+	 */
+	private $userScopeService;
+
+	private const DOCUMENT_MIME_TYPES = [
+		'document' => 'application/vnd.google-apps.document',
+		'spreadsheet' => 'application/vnd.google-apps.spreadsheet',
+		'presentation' => 'application/vnd.google-apps.presentation',
+	];
 
 	/**
 	 * Service to make requests to Google v3 (JSON) API
 	 */
 	public function __construct (string $appName,
 								LoggerInterface $logger,
-								IL10N $l10n,
 								IConfig $config,
 								IRootFolder $root,
 								IJobList $jobList,
+								UserScopeService $userScopeService,
 								GoogleAPIService $googleApiService) {
 		$this->appName = $appName;
-		$this->l10n = $l10n;
-		$this->config = $config;
 		$this->logger = $logger;
-		$this->jobList = $jobList;
+		$this->config = $config;
 		$this->root = $root;
+		$this->jobList = $jobList;
 		$this->googleApiService = $googleApiService;
+		$this->userScopeService = $userScopeService;
 	}
 
 	/**
@@ -103,12 +137,10 @@ class GoogleDriveAPIService {
 	}
 
 	/**
-	 * @param string $accessToken
 	 * @param string $userId
-	 * @param string $targetPath
 	 * @return array
 	 */
-	public function startImportDrive(string $accessToken, string $userId): array {
+	public function startImportDrive(string $userId): array {
 		$targetPath = $this->config->getUserValue($userId, Application::APP_ID, 'drive_output_dir', '/Google Drive');
 		$targetPath = $targetPath ?: '/Google Drive';
 		$alreadyImporting = $this->config->getUserValue($userId, Application::APP_ID, 'importing_drive', '0') === '1';
@@ -118,7 +150,7 @@ class GoogleDriveAPIService {
 		// create root folder
 		$userFolder = $this->root->getUserFolder($userId);
 		if (!$userFolder->nodeExists($targetPath)) {
-			$folder = $userFolder->newFolder($targetPath);
+			$userFolder->newFolder($targetPath);
 		} else {
 			$folder = $userFolder->get($targetPath);
 			if ($folder->getType() !== FileInfo::TYPE_FOLDER) {
@@ -136,10 +168,15 @@ class GoogleDriveAPIService {
 
 	/**
 	 * @param string $userId
-	 * @return array
+	 * @return void
 	 */
 	public function importDriveJob(string $userId): void {
-		$this->logger->info('Importing drive files for ' . $userId);
+		$this->logger->debug('Importing drive files for ' . $userId);
+
+		// Set the user to register the change under his name
+		$this->userScopeService->setUserScope($userId);
+		$this->userScopeService->setFilesystemScope($userId);
+
 		$importingDrive = $this->config->getUserValue($userId, Application::APP_ID, 'importing_drive', '0') === '1';
 		$jobRunning = $this->config->getUserValue($userId, Application::APP_ID, 'drive_import_running', '0') === '1';
 		if (!$importingDrive || $jobRunning) {
@@ -147,7 +184,7 @@ class GoogleDriveAPIService {
 		}
 		$this->config->setUserValue($userId, Application::APP_ID, 'drive_import_running', '1');
 
-		$accessToken = $this->config->getUserValue($userId, Application::APP_ID, 'token', '');
+		$accessToken = $this->config->getUserValue($userId, Application::APP_ID, 'token');
 		// import batch of files
 		$targetPath = $this->config->getUserValue($userId, Application::APP_ID, 'drive_output_dir', '/Google Drive');
 		$targetPath = $targetPath ?: '/Google Drive';
@@ -159,7 +196,13 @@ class GoogleDriveAPIService {
 		// import by batch of 500 Mo
 		$alreadyImported = $this->config->getUserValue($userId, Application::APP_ID, 'nb_imported_files', '0');
 		$alreadyImported = (int) $alreadyImported;
-		$result = $this->importFiles($accessToken, $userId, $targetPath, 500000000, $alreadyImported, $directoryProgress);
+		try {
+			$result = $this->importFiles($accessToken, $userId, $targetPath, 500000000, $alreadyImported, $directoryProgress);
+		} catch (\Exception | \Throwable $e) {
+			$result = [
+				'error' => 'Unknown job failure. ' . $e->getMessage(),
+			];
+		}
 		if (isset($result['error']) || (isset($result['finished']) && $result['finished'])) {
 			if (isset($result['finished']) && $result['finished']) {
 				$nbImported = (int) $this->config->getUserValue($userId, Application::APP_ID, 'nb_imported_files', '0');
@@ -177,7 +220,7 @@ class GoogleDriveAPIService {
 			$this->config->deleteUserValue($userId, Application::APP_ID, 'directory_progress');
 		} else {
 			$this->config->setUserValue($userId, Application::APP_ID, 'directory_progress', json_encode($directoryProgress));
-			$ts = (new \Datetime())->getTimestamp();
+			$ts = (new Datetime())->getTimestamp();
 			$this->config->setUserValue($userId, Application::APP_ID, 'last_drive_import_timestamp', $ts);
 			$this->jobList->add(ImportDriveJob::class, ['user_id' => $userId]);
 		}
@@ -190,10 +233,15 @@ class GoogleDriveAPIService {
 	 * @param string $targetPath
 	 * @param ?int $maxDownloadSize
 	 * @param int $alreadyImported
+	 * @param array $directoryProgress
 	 * @return array
+	 * @throws NotFoundException
+	 * @throws \OCP\Files\NotPermittedException
+	 * @throws \OCP\PreConditionNotMetException
+	 * @throws \OC\User\NoUserException
 	 */
 	public function importFiles(string $accessToken, string $userId, string $targetPath,
-								?int $maxDownloadSize = null, int $alreadyImported, array &$directoryProgress): array {
+								?int $maxDownloadSize = null, int $alreadyImported = 0, array &$directoryProgress = []): array {
 		$considerSharedFiles = $this->config->getUserValue($userId, Application::APP_ID, 'consider_shared_files', '0') === '1';
 		// create root folder
 		$userFolder = $this->root->getUserFolder($userId);
@@ -202,7 +250,7 @@ class GoogleDriveAPIService {
 		} else {
 			$folder = $userFolder->get($targetPath);
 			if ($folder->getType() !== FileInfo::TYPE_FOLDER) {
-				return ['error' => 'Impossible to create ' . $targetPath . ' folder'];
+				return ['error' => 'Impossible to create ' . '<redacted>' . ' folder'];
 			}
 		}
 
@@ -250,7 +298,7 @@ class GoogleDriveAPIService {
 		if (isset($info['error'])) {
 			return $info;
 		}
-		$nbFilesOnDrive = $info['nbFiles'];
+		// $nbFilesOnDrive = $info['nbFiles'];
 		$downloadedSize = 0;
 		$nbDownloaded = 0;
 
@@ -278,7 +326,32 @@ class GoogleDriveAPIService {
 					if (!$considerSharedFiles && !$fileItem['ownedByMe']) {
 						continue;
 					}
-					$size = $this->getFile($accessToken, $userId, $fileItem, $directoriesById, $folder);
+
+					if (isset($fileItem['parents']) && count($fileItem['parents']) > 0
+						&& isset($directoriesById[$fileItem['parents'][0]], $directoriesById[$fileItem['parents'][0]]['node'])) {
+						$saveFolder = $directoriesById[$fileItem['parents'][0]]['node'];
+					} else {
+						$saveFolder = $folder;
+					}
+
+					$fileName = $this->getFileName($fileItem, $userId);
+
+					// If file already exists in folder, don't download unless timestamp is different
+					if ($saveFolder->nodeExists($fileName)) {
+						$savedFile = $saveFolder->get($fileName);
+						$timestampOnFile = $savedFile->getMtime();
+						$d = new Datetime($fileItem['modifiedTime']);
+						$timestampOnDrive = $d->getTimestamp();
+
+						if ($timestampOnFile < $timestampOnDrive) {
+							$savedFile->delete();
+						} else {
+							continue;
+						}
+					}
+
+					$size = $this->getFile($accessToken, $userId, $fileItem, $saveFolder, $fileName);
+
 					if (!is_null($size)) {
 						$nbDownloaded++;
 						$this->config->setUserValue($userId, Application::APP_ID, 'nb_imported_files', $alreadyImported + $nbDownloaded);
@@ -290,6 +363,9 @@ class GoogleDriveAPIService {
 								'finished' => false,
 							];
 						}
+					} elseif (!$saveFolder->nodeExists($fileName)) {
+						$filePathInDrive = $dirId === 'root' ? '/' . $fileItem['name'] : $directoriesById[$dirId]['name'] . '/' . $fileItem['name'];
+						$this->logFailedDownloadsForUser($folder, $filePathInDrive);
 					}
 				}
 				$params['pageToken'] = $result['nextPageToken'] ?? '';
@@ -306,15 +382,33 @@ class GoogleDriveAPIService {
 	}
 
 	/**
+	 * @param Folder $folder
+	 * @param string $fileName
+	 * @throws LockedException
+	 * @throws \OCP\Files\NotPermittedException
+	 */
+	private function logFailedDownloadsForUser(Folder $folder, string $fileName): void {
+		try {
+			$logFile = $folder->get('failed-downloads.md');
+		} catch (NotFoundException $e) {
+			$logFile = $folder->newFile('failed-downloads.md');
+		}
+
+		$stream = $logFile->fopen('a');
+		fwrite($stream, '1. Failed to download file: ' . $fileName . PHP_EOL);
+		fclose($stream);
+	}
+
+	/**
 	 * recursive directory creation
 	 * associate the folder node to directories on the fly
 	 *
 	 * @param array &$directoriesById
-	 * @param Node $currentFolder
-	 * @param string $currentFolder
+	 * @param Folder $currentFolder
+	 * @param string $currentFolderId
 	 * @return bool success
 	 */
-	private function createDirsUnder(array &$directoriesById, Node $currentFolder, string $currentFolderId = ''): bool {
+	private function createDirsUnder(array &$directoriesById, Folder $currentFolder, string $currentFolderId = ''): bool {
 		foreach ($directoriesById as $id => $dir) {
 			$parentId = $dir['parent'];
 			// create dir if we are on top OR if its parent is current dir
@@ -340,131 +434,145 @@ class GoogleDriveAPIService {
 	}
 
 	/**
+	 * Create new file in the given folder with given filename
+	 * Download contents of the file from Google Drive and save it into the created file
+	 * @param string $accessToken
+	 * @param Folder $saveFolder
+	 * @param string $fileName
+	 * @param string $userId
+	 * @param string $fileUrl
+	 * @param array $fileItem
+	 * @param array $params
+	 * @return ?int downloaded size, null if error during file creation or download
+	 * @throws InvalidPathException
+	 * @throws LockedException
+	 * @throws NotFoundException
+	 * @throws NotPermittedException
+	 */
+	 private function downloadAndSaveFile(string $accessToken, Folder $saveFolder, string $fileName, string $userId,
+										  string $fileUrl, array $fileItem, array $params = []): ?int {
+		try {
+			$savedFile = $saveFolder->newFile($fileName);
+		} catch (NotPermittedException $e) {
+			return null;
+		}
+
+		try {
+			$resource = $savedFile->fopen('w');
+		} catch (LockedException $e) {
+			return null;
+		}
+
+		$res = $this->googleApiService->simpleDownload($accessToken, $userId, $fileUrl, $resource, $params);
+		if (!isset($res['error'])) {
+			if (is_resource($resource)) {
+				fclose($resource);
+			}
+			if (isset($fileItem['modifiedTime'])) {
+				$d = new Datetime($fileItem['modifiedTime']);
+				$ts = $d->getTimestamp();
+				$savedFile->touch($ts);
+			} else {
+				$savedFile->touch();
+			}
+			$stat = $savedFile->stat();
+			return $stat['size'] ?? 0;
+		} else {
+			if ($savedFile->isDeletable()) {
+				$savedFile->unlock(ILockingProvider::LOCK_EXCLUSIVE);
+				$savedFile->delete();
+			}
+		}
+		return null;
+	 }
+
+	/**
+	 * @param array $fileItem
+	 * @param string $userId
+	 * @return string name of the file to be saved
+	*/
+	private function getFileName(array $fileItem, string $userId): string {
+		$fileName = preg_replace('/\//', '-', $fileItem['name'] ?? 'Untitled');
+
+		if (in_array($fileItem['mimeType'], array_values(self::DOCUMENT_MIME_TYPES))) {
+			$documentFormat = $this->getUserDocumentFormat($userId);
+			switch ($fileItem['mimeType']) {
+				case self::DOCUMENT_MIME_TYPES['document']:
+					$fileName .= $documentFormat === 'openxml' ? '.docx' : '.odt';
+					break;
+				case self::DOCUMENT_MIME_TYPES['spreadsheet']:
+					$fileName .= $documentFormat === 'openxml' ? '.xlsx' : '.ods';
+					break;
+				case self::DOCUMENT_MIME_TYPES['presentation']:
+					$fileName .= $documentFormat === 'openxml' ? '.pptx' : '.odp';
+					break;
+			}
+		}
+
+		$extension = pathinfo($fileName, PATHINFO_EXTENSION);
+		$name = pathinfo($fileName, PATHINFO_FILENAME) . '_' . substr($fileItem['id'], -6);
+
+		return strlen($extension) ? $name . '.' . $extension : $name;
+	}
+
+	/**
+	 * @param string $userId
+	 * @return string User's preferred document format
+	*/
+	private function getUserDocumentFormat(string $userId): string {
+		$documentFormat = $this->config->getUserValue($userId, Application::APP_ID, 'document_format', 'openxml');
+		if (!in_array($documentFormat, ['openxml', 'opendoc'])) {
+			$documentFormat = 'openxml';
+		}
+		return $documentFormat;
+	}
+
+	/**
+	 * @param string $mimeType
+	 * @param string $documentFormat
+	 * @return array Request parameters for document
+	*/
+	private function getDocumentRequestParams(string $mimeType, string $documentFormat): array {
+		$params = [];
+		switch ($mimeType) {
+			case self::DOCUMENT_MIME_TYPES['document']:
+				$params['mimeType'] = $documentFormat === 'openxml'
+					? 'application/vnd.openxmlformats-officedocument.wordprocessingml.document'
+					: 'application/vnd.oasis.opendocument.text';
+				break;
+			case self::DOCUMENT_MIME_TYPES['spreadsheet']:
+				$params['mimeType'] = $documentFormat === 'openxml'
+					? 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+					: 'application/vnd.oasis.opendocument.spreadsheet';
+				break;
+			case self::DOCUMENT_MIME_TYPES['presentation']:
+				$params['mimeType'] = $documentFormat === 'openxml'
+					? 'application/vnd.openxmlformats-officedocument.presentationml.presentation'
+					:'application/vnd.oasis.opendocument.presentation';
+				break;
+		}
+		return $params;
+	}
+
+	/**
 	 * @param string $accessToken
 	 * @param string $userId
 	 * @param array $fileItem
-	 * @param array $directoriesById
-	 * @param Node $topFolder
-	 * @return ?int downloaded size, null if already existing
+	 * @param Folder $saveFolder
+	 * @param string $fileName
+	 * @return ?int downloaded size, null if error getting file
 	 */
-	private function getFile(string $accessToken, string $userId, array $fileItem, array $directoriesById, Node $topFolder): ?int {
-		$fileName = preg_replace('/\//', '-slash-', $fileItem['name'] ?? 'Untitled');
-		if (isset($fileItem['parents']) && count($fileItem['parents']) > 0 && array_key_exists($fileItem['parents'][0], $directoriesById)) {
-			$saveFolder = $directoriesById[$fileItem['parents'][0]]['node'];
-		} else {
-			$saveFolder = $topFolder;
-		}
-		// classic file
-		if (isset($fileItem['webContentLink'])) {
-			if (!$saveFolder->nodeExists($fileName)) {
-				$fileUrl = 'https://www.googleapis.com/drive/v3/files/' . $fileItem['id'] . '?alt=media';
-				try {
-					$savedFile = $saveFolder->newFile($fileName);
-				} catch (NotFoundException $e) {
-					$this->logger->warning(
-						'Google Drive error, can\'t create file "' . $fileName . '" in "' . $saveFolder->getPath() . '"',
-						['app' => $this->appName]
-					);
-					return null;
-				}
-				try {
-					$resource = $savedFile->fopen('w');
-				} catch (LockedException $e) {
-					$this->logger->warning('Google Drive error opening target file ' . $savedFile->getPath() . ' : file is locked', ['app' => $this->appName]);
-					return null;
-				}
-				$res = $this->googleApiService->simpleDownload($accessToken, $userId, $fileUrl, $resource);
-				if (!isset($res['error'])) {
-					if (is_resource($resource)) {
-						fclose($resource);
-					}
-					if (isset($fileItem['modifiedTime'])) {
-						$d = new \Datetime($fileItem['modifiedTime']);
-						$ts = $d->getTimestamp();
-						$savedFile->touch($ts);
-					} else {
-						$savedFile->touch();
-					}
-					$stat = $savedFile->stat();
-					return $stat['size'] ?? 0;
-				} else {
-					$this->logger->warning('Google Drive error downloading file ' . $fileItem['name'] . ' : ' . $res['error'], ['app' => $this->appName]);
-					if ($savedFile->isDeletable()) {
-						$savedFile->delete();
-					}
-				}
-			}
-		} else {
-			$documentFormat = $this->config->getUserValue($userId, Application::APP_ID, 'document_format', 'openxml');
-			if (!in_array($documentFormat, ['openxml', 'opendoc'])) {
-				$documentFormat = 'openxml';
-			}
+	private function getFile(string $accessToken, string $userId, array $fileItem, Folder $saveFolder, string $fileName): ?int {
+		if (in_array($fileItem['mimeType'], array_values(self::DOCUMENT_MIME_TYPES))) {
+			$documentFormat = $this->getUserDocumentFormat($userId);
 			// potentially a doc
-			if ($fileItem['mimeType'] === 'application/vnd.google-apps.document') {
-				$fileName .= $documentFormat === 'openxml' ? '.docx' : '.odt';
-				$mimeType = $documentFormat === 'openxml'
-					? 'application/vnd.openxmlformats-officedocument.wordprocessingml.document'
-					: 'application/vnd.oasis.opendocument.text';
-			} elseif ($fileItem['mimeType'] === 'application/vnd.google-apps.spreadsheet') {
-				$fileName .= $documentFormat === 'openxml' ? '.xlsx' : '.ods';
-				$mimeType = $documentFormat === 'openxml'
-					? 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
-					: 'application/vnd.oasis.opendocument.spreadsheet';
-			} elseif ($fileItem['mimeType'] === 'application/vnd.google-apps.presentation') {
-				$fileName .= $documentFormat === 'openxml' ? '.pptx' : '.odp';
-				$mimeType = $documentFormat === 'openxml'
-					? 'application/vnd.openxmlformats-officedocument.presentationml.presentation'
-					:'application/vnd.oasis.opendocument.presentation';
-			} else {
-				$this->logger->warning(
-					'Google Drive error downloading file, no webContentLink, unknown mime type: ' . $saveFolder->getInternalPath() . '/' . ($fileItem['name'] ?? 'Untitled') . ' : '
-						. json_encode($fileItem),
-					['app' => $this->appName]
-				);
-				return null;
-			}
-			if (!$saveFolder->nodeExists($fileName)) {
-				$params = [
-					'mimeType' => $mimeType,
-				];
-				$fileUrl = 'https://www.googleapis.com/drive/v3/files/' . $fileItem['id'] . '/export';
-				$savedFile = $saveFolder->newFile($fileName);
-				try {
-					$savedFile = $saveFolder->newFile($fileName);
-				} catch (NotFoundException $e) {
-					$this->logger->warning(
-						'Google Drive error, can\'t create document file "' . $fileName . '" in "' . $saveFolder->getPath() . '"',
-						['app' => $this->appName]
-					);
-					return null;
-				}
-				try {
-					$resource = $savedFile->fopen('w');
-				} catch (LockedException $e) {
-					$this->logger->warning('Google Drive error opening target file ' . $savedFile->getPath() . ' : file is locked', ['app' => $this->appName]);
-					return null;
-				}
-				$res = $this->googleApiService->simpleDownload($accessToken, $userId, $fileUrl, $resource, $params);
-				if (!isset($res['error'])) {
-					if (is_resource($resource)) {
-						fclose($resource);
-					}
-					if (isset($fileItem['modifiedTime'])) {
-						$d = new \Datetime($fileItem['modifiedTime']);
-						$ts = $d->getTimestamp();
-						$savedFile->touch($ts);
-					} else {
-						$savedFile->touch();
-					}
-					$stat = $savedFile->stat();
-					return $stat['size'] ?? 0;
-				} else {
-					$this->logger->warning('Google Drive error downloading file ' . $fileItem['name'] . ' : ' . $res['error'], ['app' => $this->appName]);
-					if ($savedFile->isDeletable()) {
-						$savedFile->delete();
-					}
-				}
-			}
+			$params = $this->getDocumentRequestParams($fileItem['mimeType'], $documentFormat);
+			$fileUrl = 'https://www.googleapis.com/drive/v3/files/' . urlencode($fileItem['id']) . '/export';
+			return $this->downloadAndSaveFile($accessToken, $saveFolder, $fileName, $userId, $fileUrl, $fileItem, $params);
+		} elseif (isset($fileItem['webContentLink'])) {
+			// classic file
+			$fileUrl = 'https://www.googleapis.com/drive/v3/files/' . urlencode($fileItem['id']) . '?alt=media';
+			return $this->downloadAndSaveFile($accessToken, $saveFolder, $fileName, $userId, $fileUrl, $fileItem);
 		}
 		return null;
 	}