+++ lib/Service/GooglePhotosAPIService.php
@@ -11,11 +11,12 @@
 
 namespace OCA\Google\Service;
 
-use OCP\IL10N;
+use Datetime;
+use OCP\Files\Folder;
 use OCP\IConfig;
 use OCP\Files\IRootFolder;
 use OCP\Files\FileInfo;
-use OCP\Files\Node;
+use OCP\Lock\ILockingProvider;
 use OCP\Lock\LockedException;
 use OCP\BackgroundJob\IJobList;
 use Psr\Log\LoggerInterface;
@@ -24,27 +25,52 @@ use OCA\Google\AppInfo\Application;
 use OCA\Google\BackgroundJob\ImportPhotosJob;
 
 class GooglePhotosAPIService {
-
-	private $l10n;
+	/**
+	 * @var string
+	 */
+	private $appName;
+	/**
+	 * @var LoggerInterface
+	 */
 	private $logger;
+	/**
+	 * @var IConfig
+	 */
+	private $config;
+	/**
+	 * @var IRootFolder
+	 */
+	private $root;
+	/**
+	 * @var IJobList
+	 */
+	private $jobList;
+	/**
+	 * @var GoogleAPIService
+	 */
+	private $googleApiService;
+	/**
+	 * @var UserScopeService
+	 */
+	private $userScopeService;
 
 	/**
 	 * Service to make requests to Google v3 (JSON) API
 	 */
 	public function __construct (string $appName,
 								LoggerInterface $logger,
-								IL10N $l10n,
 								IConfig $config,
 								IRootFolder $root,
 								IJobList $jobList,
+								UserScopeService $userScopeService,
 								GoogleAPIService $googleApiService) {
 		$this->appName = $appName;
-		$this->l10n = $l10n;
-		$this->config = $config;
 		$this->logger = $logger;
-		$this->jobList = $jobList;
+		$this->config = $config;
 		$this->root = $root;
+		$this->jobList = $jobList;
 		$this->googleApiService = $googleApiService;
+		$this->userScopeService = $userScopeService;
 	}
 
 	/**
@@ -58,7 +84,7 @@ class GooglePhotosAPIService {
 			'pageSize' => 50,
 		];
 		do {
-			$this->logger->info(
+			$this->logger->debug(
 				'Photos service::getPhotoNumber LAUNCHING ALBUM LIST REQUEST, userid: "' . $userId . '", token length: ' . strlen($accessToken),
 				['app' => $this->appName]
 			);
@@ -70,12 +96,6 @@ class GooglePhotosAPIService {
 				foreach ($result['albums'] as $album) {
 					$nbPhotos += $album['mediaItemsCount'] ?? 0;
 				}
-			} else {
-				$this->logger->warning(
-					'Google API error getting album list to get photo number, no "albums" key in '
-						. json_encode($result),
-					['app' => $this->appName]
-				);
 			}
 			$params['pageToken'] = $result['nextPageToken'] ?? '';
 		} while (isset($result['nextPageToken']));
@@ -83,7 +103,6 @@ class GooglePhotosAPIService {
 		// shared albums
 		$considerSharedAlbums = $this->config->getUserValue($userId, Application::APP_ID, 'consider_shared_albums', '0') === '1';
 		if ($considerSharedAlbums) {
-			$sharedAlbums = [];
 			$params = [
 				'pageSize' => 50,
 			];
@@ -96,34 +115,50 @@ class GooglePhotosAPIService {
 					foreach ($result['sharedAlbums'] as $album) {
 						$nbPhotos += $album['mediaItemsCount'] ?? 0;
 					}
-				} else {
-					$this->logger->warning(
-						'Google API error getting shared albums list to get photo number, no "sharedAlbums" key in '
-							. json_encode($result),
-						['app' => $this->appName]
-					);
 				}
 				$params['pageToken'] = $result['nextPageToken'] ?? '';
 			} while (isset($result['nextPageToken']));
 		}
+
+		// check if there is any photo outside albums
+		// (number is not relevant here as we just make one paginated request to avoid reaching request limit)
+		if ($nbPhotos === 0) {
+			$params = [
+				'pageSize' => 50,
+			];
+
+			$result = $this->googleApiService->request($accessToken, $userId, 'v1/mediaItems', $params, 'GET', 'https://photoslibrary.googleapis.com/');
+			if (isset($result['error'])) {
+				return $result;
+			}
+
+			if (isset($result['mediaItems']) && is_array($result['mediaItems'])) {
+				$nbPhotos += count($result['mediaItems']);
+			} else {
+				$this->logger->warning(
+					'Google API error getting media items list to get photo number, no "mediaItems" key in '
+					. json_encode($result),
+					['app' => $this->appName]
+				);
+			}
+		}
+
 		return [
 			'nbPhotos' => $nbPhotos,
 		];
 	}
 
 	/**
-	 * @param string $accessToken
 	 * @param string $userId
-	 * @param string $targetPath
 	 * @return array
 	 */
-	public function startImportPhotos(string $accessToken, string $userId): array {
+	public function startImportPhotos(string $userId): array {
 		$targetPath = $this->config->getUserValue($userId, Application::APP_ID, 'photo_output_dir', '/Google Photos');
 		$targetPath = $targetPath ?: '/Google Photos';
 		// create root folder
 		$userFolder = $this->root->getUserFolder($userId);
 		if (!$userFolder->nodeExists($targetPath)) {
-			$folder = $userFolder->newFolder($targetPath);
+			$userFolder->newFolder($targetPath);
 		} else {
 			$folder = $userFolder->get($targetPath);
 			if ($folder->getType() !== FileInfo::TYPE_FOLDER) {
@@ -140,10 +175,15 @@ class GooglePhotosAPIService {
 
 	/**
 	 * @param string $userId
-	 * @return array
+	 * @return void
 	 */
 	public function importPhotosJob(string $userId): void {
-		$this->logger->info('Importing photos for ' . $userId);
+		$this->logger->debug('Importing photos for ' . $userId);
+
+		// Set the user to register the change under his name
+		$this->userScopeService->setUserScope($userId);
+		$this->userScopeService->setFilesystemScope($userId);
+
 		$importingPhotos = $this->config->getUserValue($userId, Application::APP_ID, 'importing_photos', '0') === '1';
 		$jobRunning = $this->config->getUserValue($userId, Application::APP_ID, 'photo_import_running', '0') === '1';
 		if (!$importingPhotos || $jobRunning) {
@@ -151,13 +191,19 @@ class GooglePhotosAPIService {
 		}
 		$this->config->setUserValue($userId, Application::APP_ID, 'photo_import_running', '1');
 
-		$accessToken = $this->config->getUserValue($userId, Application::APP_ID, 'token', '');
+		$accessToken = $this->config->getUserValue($userId, Application::APP_ID, 'token');
 		$targetPath = $this->config->getUserValue($userId, Application::APP_ID, 'photo_output_dir', '/Google Photos');
 		$targetPath = $targetPath ?: '/Google Photos';
 		// import photos by batch of 500 Mo
 		$alreadyImported = $this->config->getUserValue($userId, Application::APP_ID, 'nb_imported_photos', '0');
 		$alreadyImported = (int) $alreadyImported;
-		$result = $this->importPhotos($accessToken, $userId, $targetPath, 500000000, $alreadyImported);
+		try {
+			$result = $this->importPhotos($accessToken, $userId, $targetPath, 500000000, $alreadyImported);
+		} catch (\Exception | \Throwable $e) {
+			$result = [
+				'error' => 'Unknown job failure. ' . $e->getMessage(),
+			];
+		}
 		if (isset($result['error']) || (isset($result['finished']) && $result['finished'])) {
 			$this->config->setUserValue($userId, Application::APP_ID, 'importing_photos', '0');
 			$this->config->setUserValue($userId, Application::APP_ID, 'nb_imported_photos', '0');
@@ -172,7 +218,7 @@ class GooglePhotosAPIService {
 				$this->logger->error('Google Photo import error: ' . $result['error'], ['app' => $this->appName]);
 			}
 		} else {
-			$ts = (new \Datetime())->getTimestamp();
+			$ts = (new Datetime())->getTimestamp();
 			$this->config->setUserValue($userId, Application::APP_ID, 'last_import_timestamp', $ts);
 			$this->jobList->add(ImportPhotosJob::class, ['user_id' => $userId]);
 		}
@@ -188,7 +234,7 @@ class GooglePhotosAPIService {
 	 * @return array
 	 */
 	public function importPhotos(string $accessToken, string $userId, string $targetPath,
-								?int $maxDownloadSize = null, int $alreadyImported): array {
+								?int $maxDownloadSize = null, int $alreadyImported = 0): array {
 		// create root folder
 		$userFolder = $this->root->getUserFolder($userId);
 		if (!$userFolder->nodeExists($targetPath)) {
@@ -205,7 +251,7 @@ class GooglePhotosAPIService {
 			'pageSize' => 50,
 		];
 		do {
-			$this->logger->info(
+			$this->logger->debug(
 				'Photos service::importPhotos LAUNCHING ALBUM LIST REQUEST, userid: "' . $userId . '", token length: ' . strlen($accessToken),
 				['app' => $this->appName]
 			);
@@ -217,12 +263,6 @@ class GooglePhotosAPIService {
 				foreach ($result['albums'] as $album) {
 					$albums[] = $album;
 				}
-			} else {
-				$this->logger->warning(
-					'Google API error getting album list when importing, no "albums" key in '
-						. json_encode($result),
-					['app' => $this->appName]
-				);
 			}
 			$params['pageToken'] = $result['nextPageToken'] ?? '';
 		} while (isset($result['nextPageToken']));
@@ -242,15 +282,13 @@ class GooglePhotosAPIService {
 					foreach ($result['sharedAlbums'] as $album) {
 						$albums[] = $album;
 					}
-				} else {
-					$this->logger->warning('Google API error getting shared albums list, no "sharedAlbums" key in ' . json_encode($result), ['app' => $this->appName]);
 				}
 				$params['pageToken'] = $result['nextPageToken'] ?? '';
 			} while (isset($result['nextPageToken']));
 		}
 
 		// get the photos
-		$this->logger->info(
+		$this->logger->debug(
 			'Photos service::importPhotos GETTING PHOTOS, nb albums: "' . count($albums) . '"',
 			['app' => $this->appName]
 		);
@@ -298,8 +336,6 @@ class GooglePhotosAPIService {
 							}
 						}
 					}
-				} else {
-					$this->logger->warning('Google API error getting photo list, no "mediaItems" key in ' . json_encode($result), ['app' => $this->appName]);
 				}
 				$params['pageToken'] = $result['nextPageToken'] ?? '';
 			} while (isset($result['nextPageToken']));
@@ -335,8 +371,6 @@ class GooglePhotosAPIService {
 						}
 					}
 				}
-			} else {
-				$this->logger->warning('Google API error getting photo list, no "mediaItems" key in ' . json_encode($result), ['app' => $this->appName]);
 			}
 			$params['pageToken'] = $result['nextPageToken'] ?? '';
 		} while (isset($result['nextPageToken']));
@@ -353,10 +387,13 @@ class GooglePhotosAPIService {
 	 * @param string $accessToken
 	 * @param string $userId
 	 * @param array $photo
-	 * @param Node $albumFolder
+	 * @param Folder $albumFolder
 	 * @return ?int downloaded size, null if already existing
+	 * @throws \OCP\Files\InvalidPathException
+	 * @throws \OCP\Files\NotFoundException
+	 * @throws \OCP\Files\NotPermittedException
 	 */
-	private function getPhoto(string $accessToken, string $userId, array $photo, Node $albumFolder): ?int {
+	private function getPhoto(string $accessToken, string $userId, array $photo, Folder $albumFolder): ?int {
 		$photoName = $photo['filename'];
 		if (!$albumFolder->nodeExists($photoName)) {
 			if (isset($photo['mediaMetadata']['photo'])) {
@@ -373,7 +410,7 @@ class GooglePhotosAPIService {
 			try {
 				$resource = $savedFile->fopen('w');
 			} catch (LockedException $e) {
-				$this->logger->warning('Google Photo, error opening target file ' . $savedFile->getPath() . ' : file is locked', ['app' => $this->appName]);
+				$this->logger->warning('Google Photo, error opening target file ' . '<redacted>' . ' : file is locked', ['app' => $this->appName]);
 				return null;
 			}
 			$res = $this->googleApiService->simpleDownload($accessToken, $userId, $photoUrl, $resource);
@@ -382,7 +419,7 @@ class GooglePhotosAPIService {
 					fclose($resource);
 				}
 				if (isset($photo['mediaMetadata']['creationTime'])) {
-					$d = new \Datetime($photo['mediaMetadata']['creationTime']);
+					$d = new Datetime($photo['mediaMetadata']['creationTime']);
 					$ts = $d->getTimestamp();
 					$savedFile->touch($ts);
 				} else {
@@ -391,8 +428,9 @@ class GooglePhotosAPIService {
 				$stat = $savedFile->stat();
 				return $stat['size'] ?? 0;
 			} else {
-				$this->logger->warning('Google API error downloading photo ' . $photoName . ' : ' . $res['error'], ['app' => $this->appName]);
+				$this->logger->warning('Google API error downloading photo ' . '<redacted>' . ' : ' . $res['error'], ['app' => $this->appName]);
 				if ($savedFile->isDeletable()) {
+					$savedFile->unlock(ILockingProvider::LOCK_EXCLUSIVE);
 					$savedFile->delete();
 				}
 			}